# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ExecFleet API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/company": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create a new company",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Company"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Company added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "logo_url": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "address": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "string"
                    },
                    "contact_person": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Company with this login email already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get all companies",
        "responses": {
          "200": {
            "description": "List of companies"
          }
        }
      }
    },
    "/api/admin/company/{company_id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get company by ID",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company found"
          },
          "404": {
            "description": "Company not found"
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Update company by ID",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Company"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Company updated"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/admin/company": {
      "post": {
        "summary": "Create a new company",
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "logo": {
                    "type": "string",
                    "format": "binary"
                  },
                  "lat": {
                    "type": "number"
                  },
                  "lng": {
                    "type": "number"
                  },
                  "bussiness_nature": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "number_of_employees": {
                    "type": "number"
                  },
                  "address": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "state": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "contact_person": {
                    "type": "string"
                  },
                  "login_email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "logo",
                  "name",
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Company added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "logo_url": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "address": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "string"
                    },
                    "contact_person": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an existing company",
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "company_id": {
                    "type": "string"
                  },
                  "logo": {
                    "type": "string",
                    "format": "binary"
                  },
                  "lat": {
                    "type": "number"
                  },
                  "lng": {
                    "type": "number"
                  },
                  "bussiness_nature": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "number_of_employees": {
                    "type": "number"
                  },
                  "address": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "state": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "contact_person": {
                    "type": "string"
                  },
                  "login_email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "company_id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Company updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "logo_url": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "address": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "string"
                    },
                    "contact_person": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List companies (with pagination and search)",
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Search by company name"
          }
        ],
        "responses": {
          "200": {
            "description": "List of companies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "address": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "phone": {
                        "type": "string"
                      },
                      "contact_person": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/company/{company_id}": {
      "get": {
        "tags": [
          "Admin",
          "Company"
        ],
        "summary": "Get company by ID",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company found"
          },
          "404": {
            "description": "Company not found"
          }
        }
      },
      "put": {
        "tags": [
          "Admin",
          "Company"
        ],
        "summary": "Update company by ID",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Company updated"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/api/admin/route": {
      "post": {
        "tags": [
          "Admin",
          "Routes"
        ],
        "summary": "Create a new route",
        "description": "Creates a new route and links it with a driver and optionally their vehicle.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Morning Route"
                  },
                  "routes": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "employee_id": {
                          "type": "string",
                          "description": "MongoDB ObjectId of the employee",
                          "example": "64cabc1234567890abcdef12"
                        },
                        "address": {
                          "type": "string",
                          "example": "123 Main Street"
                        },
                        "location": {
                          "type": "object",
                          "properties": {
                            "lat": {
                              "type": "number",
                              "example": 24.8607
                            },
                            "lng": {
                              "type": "number",
                              "example": 67.0011
                            }
                          }
                        }
                      },
                      "required": [
                        "employee_id",
                        "address",
                        "location"
                      ]
                    }
                  },
                  "city": {
                    "type": "string",
                    "example": "Karachi"
                  },
                  "from": {
                    "type": "string",
                    "example": "Office"
                  },
                  "to": {
                    "type": "string",
                    "example": "Client Site"
                  },
                  "kilometers": {
                    "type": "number",
                    "example": 12.5
                  },
                  "driver_id": {
                    "type": "string",
                    "description": "MongoDB ObjectId of the driver",
                    "example": "64df1234567890abcdef12"
                  }
                },
                "required": [
                  "name",
                  "routes",
                  "city",
                  "from",
                  "to",
                  "kilometers",
                  "driver_id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Route created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "64cabcd9876543210fedcba1"
                    },
                    "name": {
                      "type": "string",
                      "example": "Morning Route"
                    },
                    "city": {
                      "type": "string",
                      "example": "Karachi"
                    },
                    "from": {
                      "type": "string",
                      "example": "Office"
                    },
                    "to": {
                      "type": "string",
                      "example": "Client Site"
                    },
                    "kilometers": {
                      "type": "number",
                      "example": 12.5
                    },
                    "driver_name": {
                      "type": "string",
                      "example": "Ali Raza"
                    },
                    "driver_phone": {
                      "type": "string",
                      "example": "03001234567"
                    },
                    "vehicle_number": {
                      "type": "string",
                      "example": "ABC-123"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "404": {
            "description": "Driver or related resource not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/admin/routes": {
      "get": {
        "tags": [
          "Admin - Routes"
        ],
        "summary": "Get all routes",
        "description": "Returns a list of all saved routes with driver and vehicle details.",
        "responses": {
          "200": {
            "description": "List of routes retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "60f71bf75fc13c001c9e4c25"
                      },
                      "name": {
                        "type": "string",
                        "example": "Route A"
                      },
                      "city": {
                        "type": "string",
                        "example": "Lahore"
                      },
                      "from": {
                        "type": "string",
                        "example": "Gulberg"
                      },
                      "to": {
                        "type": "string",
                        "example": "DHA"
                      },
                      "kilometers": {
                        "type": "number",
                        "example": 12.5
                      },
                      "driver_name": {
                        "type": "string",
                        "example": "Ali Raza"
                      },
                      "driver_phone": {
                        "type": "string",
                        "example": "03001234567"
                      },
                      "vehicle_number": {
                        "type": "string",
                        "example": "ABC-123"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error while fetching routes"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Company": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "contact_person": {
            "type": "string"
          },
          "logo_url": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "bussiness_nature": {
            "type": "string"
          },
          "number_of_employees": {
            "type": "number"
          },
          "lat": {
            "type": "number"
          },
          "lng": {
            "type": "number"
          }
        }
      },
...
    }
  }
}